Run integration of human (manton) and canine data
This was the first sucessful integration using the Seurat
v4 integration approach. Due to previous errors, this approach
had to exclude a substantial portion of the Manton dataset
to accomidate limitations of sprase matricies.

NOTE: the allCellslabTransfer_S3.rds object loaded in for
this run was overridden and was a later object that had
some additional cells filtered out.

TO DO: re-run with appropriate allCellslabTransfer_S3.rds
object.

JOBID: 5216321