Add pseudotime ploting using moving avg
Monocle 3's plot_genes_in_pseudotime function does not appear
to be functioning properly. This is a custom apporach to plot
a similar data visual.

The apporach here uses a moving average to plot GEX over pseudo-
time. The moving average is calcualted using bins of 5 cells across
a rescaled pseudotime that ranges from 0 to 1.

In contrast, plot_genes_in_pseudotime would use a generalized linear model
to fit the data, but there is a problem with their function and I
think that moving average does a good job of converying the key point.

The code added in this commit will be split out into a seperate script.